+++ src/otf_api/api/api.py
@@ -5,7 +5,7 @@ from .bookings import BookingApi
 from .client import OtfClient
 from .members import MemberApi
 from .studios import StudioApi
-from .trends import TrendsApi
+from .trends import TrendApi
 from .workouts import WorkoutApi
 
 
@@ -25,7 +25,7 @@ class Otf:
     members: MemberApi
     workouts: WorkoutApi
     studios: StudioApi
-    trends: TrendsApi
+    trends: TrendApi
 
     def __init__(self, user: OtfUser | None = None):
         """Initialize the OTF API client.
@@ -39,7 +39,7 @@ class Otf:
         self.members = MemberApi(self, client)
         self.workouts = WorkoutApi(self, client)
         self.studios = StudioApi(self, client)
-        self.trends = TrendsApi(self, client)
+        self.trends = TrendApi(self, client)
 
         self._member: models.MemberDetail | None = None
