+++ src/otf_api/models/trends/__init__.py
@@ -1,8 +1,7 @@
+from .enums import TrendCategory, TrendType
 from .trends import (
     PreviewStat,
     StatPoint,
-    TrendCategory,
-    TrendType,
     WorkoutStatsPreviewResponse,
     WorkoutStatsResponse,
 )
